refactor & merge indiv_spec and saa descriptions
