+++ scousepy/indiv_spec_description.py
@@ -10,80 +10,14 @@ CONTACT: henshaw@mpia.de
 
 import numpy as np
 from .saa_description import get_rms
+from .base_spectrum import BaseSpectrum
 
-class spectrum(object):
-    def __init__(self, coords, flux, \
-                 idx=None, scouse=None):
-        """
-        Stores all the information regarding individual spectral averaging areas
+class spectrum(BaseSpectrum):
+    
+    def __init__(self, *args, **kwargs):
+        super(self, BaseSpectrum).__init__(*args, **kwargs)
 
-        """
-
-        self._index = idx
-        self._coordinates = np.array(coords, dtype='int')
-        self._rms = get_rms(self, scouse, flux)
-        self._model_parent = None
-        self._model_spatial = None
-        self._model_dud = None
-        self._models = None
-        self._model = None
         self._decision = 'original'
-
-    @property
-    def index(self):
-        """
-        Returns the index of the spectral averaging area.
-        """
-        return self._index
-
-    @property
-    def coordinates(self):
-        """
-        Returns the coordinates of the spectral averaging area.
-        """
-        return self._coordinates
-
-    @property
-    def rms(self):
-        """
-        Returns the spectral rms.
-        """
-        return self._rms
-
-    @property
-    def model_parent(self):
-        """
-        Returns the best-fitting model derived from the parent SAA model.
-        """
-        return self._model_parent
-
-    @property
-    def model_spatial(self):
-        """
-        Returns the best-fitting model which incorporates spatial fitting.
-        """
-        return self._model_spatial
-
-    @property
-    def model_dud(self):
-        """
-        Returns the dud
-        """
-        return self._model_dud
-
-    @property
-    def models(self):
-        """
-        Returns the list of available models
-        """
-        return self._models
-
-    @property
-    def model(self):
-        """
-        Returns the best-fitting model model to the data
-        """
-        return self._model
 
     @property
     def decision(self):
